+++ src/components/Button/Button.tsx
@@ -1,10 +1,8 @@
 import React from 'react';
-import { SemanticColorToken } from '../../utils/new-colors';
 import { Icon } from '../Icon/Icon'; // Deprecated
 import { IconId } from '../../utils/icon-list'; // Deprecated
 import { Color } from '../../utils/colors'; //Deprecated
 import { ButtonBase, ButtonBaseProps } from '../ButtonBase/ButtonBase';
-import { GradeBand } from 'src/enum/gradeband';
 
 // Existin props no part of ButtonBase. To be removed in v2.0.0.
 type ExistingButtonProps = {
@@ -60,25 +58,13 @@ type ExistingButtonProps = {
 
 // Define props specific to the new button implementation
 type SpecificButtonProps = {
-  variant?: 'text' | 'contained' | 'outlined' | 'plain';
-  state?: 'activated' | 'visited';
-  color?: SemanticColorToken;
-  size?: 'sm' | 'md';
-  disableElevation?: boolean;
   fullWidth?: boolean;
-  iconOpacity?: number;
-  gradeBand?: GradeBand;
 };
 
 export type ButtonProps = ButtonBaseProps & SpecificButtonProps & ExistingButtonProps;
 
 export const Button: React.FC<ButtonProps> = ({
   children,
-  variant,
-  state,
-  color,
-  size = 'md',
-  disableElevation = false,
   fullWidth = false,
   onClick,
   disabled,
@@ -101,14 +87,8 @@ export const Button: React.FC<ButtonProps> = ({
   ...props
 }) => {
   const classNames = [
-    'connect__button',
-    variant && `connect__button-${variant}`,
-    state && `connect__button-${state}`,
-
-    color && `connect__button-${color}`,
-    size === 'sm' && 'connect__button-small',
-    disableElevation && 'connect__button-no-elevation',
     fullWidth && 'connect__button-full-width',
+    classes,
 
     // // To be removed in v2.0.0
     primary !== undefined
@@ -121,7 +101,6 @@ export const Button: React.FC<ButtonProps> = ({
     incorrect && 'connect__button-negative', // Deprecated
     mediaButton && 'connect__button-media', // Deprecated
     additionalClasses, // Deprecated
-    classes,
   ]
     .filter(Boolean)
     .join(' ');